pyproject.toml
[tool.poetry]
name = "llm-non-causal-ner"
version = "0.1.0"
description = ""
authors = ["Fred-Eric Kirsi <vegetablejuiceftw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.44.2"
datasets = "^3.0.0"
torch = "^2.4.1"
seqeval = "^1.2.2"
accelerate = "^0.34.2"
optimum = "^1.22.0"
flash-attn = "^2.6.3"
billm = "^0.1.6"
peft = "^0.12.0"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}
evaluate = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
